fix: resolve some require cycles
