+++ src/module/auth/AuthNavigatorGroup.types.ts
@@ -1,0 +1,7 @@
+export enum AuthScreens {
+    AUTH_SWITCH,
+    IMPORT_WALLET,
+    CREATE_WALLET,
+    WELCOME_BACK,
+    INTRODUCE_PIN,
+}
