+++ src/module/auth/screen/WelcomeBackScreen/WelcomeBackScreen.tsx
@@ -1,10 +1,10 @@
 import { useEffect } from "react";
 import { Typography, useSetTab } from "@peersyst/react-native-components";
 import { WelcomeBackScreenRoot } from "./WelcomeBackScreen.styles";
-import { AuthScreens } from "module/auth/AuthNavigatorGroup";
 import { useLogoPageFlex, useLogoPageGradient } from "module/common/component/layout/LogoPage/LogoPageContext";
 import { useTranslate } from "module/common/hook/useTranslate";
 import DarkThemeProvider from "module/common/component/util/ThemeProvider/DarkThemeProvider";
+import { AuthScreens } from "module/auth/AuthNavigatorGroup.types";
 
 const WelcomeBackScreen = (): JSX.Element => {
     const setTab = useSetTab();
