+++ src/module/staking/component/core/AddStakeModal/AddStakeModal.types.ts
@@ -1,0 +1,6 @@
+export enum AddStakeScreens {
+    SET_AMOUNT,
+    SELECT_VALIDATOR,
+    CONFIRM_VALIDATOR,
+    SUCCESS,
+}
