+++ src/module/staking/component/core/WithdrawModal/WithdrawModal.types.ts
@@ -1,0 +1,5 @@
+export enum WithdrawModalScreens {
+    SELECT_VALIDATOR,
+    CONFIRM_VALIDATOR,
+    SUCCESS,
+}
