+++ src/module/staking/screen/SelectValidatorScreen/AddStakeSelectValidatorScreen/AddStakeSelectValidatorScreen.tsx
@@ -1,8 +1,8 @@
 import { useTranslate } from "module/common/hook/useTranslate";
 import useGetAllValidators from "module/staking/query/useGetAllValidators";
 import { useSetTab } from "@peersyst/react-native-components";
-import { AddStakeScreens } from "module/staking/component/core/AddStakeModal/AddStakeModal";
 import StakingSelectValidatorScreen from "../StakingSelectValidatorScreen";
+import { AddStakeScreens } from "module/staking/component/core/AddStakeModal/AddStakeModal.types";
 
 const AddStakeSelectValidatorScreen = (): JSX.Element => {
     const translate = useTranslate();
