+++ src/module/staking/screen/SelectValidatorScreen/WithdrawSelectValidatorScreen/WithdrawSelectValidatorScreen.tsx
@@ -1,8 +1,8 @@
 import { useTranslate } from "module/common/hook/useTranslate";
 import { useSetTab } from "@peersyst/react-native-components";
-import { WithdrawModalScreens } from "module/staking/component/core/WithdrawModal/WithdrawModal";
 import useGetWithdrawValidators from "module/staking/hook/useGetWithdrawValidators";
 import StakingSelectValidatorScreen from "module/staking/screen/SelectValidatorScreen/StakingSelectValidatorScreen";
+import { WithdrawModalScreens } from "module/staking/component/core/WithdrawModal/WithdrawModal.types";
 
 const WithdrawSelectValidatorScreen = (): JSX.Element => {
     const translate = useTranslate();
