+++ src/module/staking/screen/SetAmountScreen/AddStakeSetAmountScreen/AddStakeSetAmountScreen.tsx
@@ -1,7 +1,7 @@
 import { useSetTab } from "@peersyst/react-native-components";
 import { useTranslate } from "module/common/hook/useTranslate";
-import { AddStakeScreens } from "module/staking/component/core/AddStakeModal/AddStakeModal";
 import StakingSetAmountScreen from "../StakingSetAmountScreen";
+import { AddStakeScreens } from "module/staking/component/core/AddStakeModal/AddStakeModal.types";
 
 const AddStakeSetAmountScreen = () => {
     const translate = useTranslate();
