+++ src/module/transaction/screen/SendScreens.types.ts
@@ -1,0 +1,5 @@
+export enum SendScreens {
+    SEND_TO_ADDRESS,
+    AMOUNT_AND_MESSAGE,
+    CONFIRMATION,
+}
