+++ src/module/transaction/screen/SendSetAmountScreen/SendSetAmountScreen.tsx
@@ -3,7 +3,6 @@ import Button from "module/common/component/input/Button/Button";
 import { useRecoilState } from "recoil";
 import sendRecoilState, { SendState } from "module/transaction/state/SendState";
 import useGetBalance from "module/wallet/query/useGetBalance";
-import { SendScreens } from "module/transaction/component/core/SendModal/SendModal";
 import CenteredLoader from "module/common/component/feedback/CenteredLoader/CenteredLoader";
 import { useTranslate } from "module/common/hook/useTranslate";
 import WalletAssetSelect from "module/wallet/component/input/WalletAssetSelect/WalletAssetSelect";
@@ -11,6 +10,7 @@ import AssetAmountTextField from "module/transaction/component/input/AssetAmount
 import { useState } from "react";
 import { Asset } from "module/wallet/wallet.types";
 import { AssetType } from "module/wallet/wallet.types";
+import { SendScreens } from "../SendScreens.types";
 
 export type SendAmountAndMessageResult = Pick<SendState, "amount" | "asset">;
