+++ test/unit/src/module/staking/screen/SetAmountScreen/AddStakeSetAmountScreen/AddStakeSetAmountScreen.spec.tsx
@@ -1,5 +1,5 @@
 import { config } from "config";
-import { AddStakeScreens } from "module/staking/component/core/AddStakeModal/AddStakeModal";
+import { AddStakeScreens } from "module/staking/component/core/AddStakeModal/AddStakeModal.types";
 import AddStakeSetAmountScreen from "module/staking/screen/SetAmountScreen/AddStakeSetAmountScreen/AddStakeSetAmountScreen";
 import { ACTION_LABEL } from "module/wallet/component/display/Balance/constants/actionLabels";
 import { CURRENCY_UNIT } from "module/wallet/component/display/Balance/constants/currencies";
